lambdas/orchestrators/requirements.txt
boto3>=1.26.0
python-dotenv>=1.0.0
langchain>=0.1.7
langchain-core>=0.1.22
langchain-anthropic>=0.1.1
langgraph>=0.0.23
